split distance.py into 3 separate modules
